removing the antd libraryu
