+++ src/Components/Common/NewCommentForm.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react';
-import 'antd/dist/antd.css';
-import Form from 'antd/es/form'
-import Input from 'antd/es/input'
-import Button from 'antd/es/button'
+// import 'antd/dist/antd.css';
+// import Form from 'antd/es/form'
+// import Input from 'antd/es/input'
+// import Button from 'antd/es/button'
 import { useGlobal } from '../../state';
 
 const layout = {
@@ -22,43 +22,44 @@ const tailLayout = {
 
 
 const NewCommentForm = props => {
-  const [ global ] = useGlobal()
-  const [form] = Form.useForm();
+  // const [ global ] = useGlobal()
+  // const [form] = Form.useForm();
 
-  useEffect(() => {
-    form.resetFields()
-  }, [props.resetSwitch])
+  // useEffect(() => {
+  //   form.resetFields()
+  // }, [props.resetSwitch])
 
-  const onFinish = formData => {
-    props.onSubmit({comment: formData.comment}, props.page)
-  }
+  // const onFinish = formData => {
+  //   props.onSubmit({comment: formData.comment}, props.page)
+  // }
 
-  const onFinishFailed = errorInfo => {} // this function has to be defined
+  // const onFinishFailed = errorInfo => {} // this function has to be defined
 
-  return (
-    <Form
-    form={form}
-      {...layout}
-      name="basic"
-      initialValues={{}}
-      onFinish={onFinish}
-      onFinishFailed={onFinishFailed}
-      style={{paddingTop: '50px', display: props.display}}
-    >
-      <Form.Item name="comment" rules={[]}>
-        <Input.TextArea />
-      </Form.Item>
-      <Form.Item {...tailLayout}>
-        <Button type="secondary" onClick={()=> form.resetFields()}>
-          Cancel
-        </Button>
-        <Button type="primary" htmlType="submit" disabled={props.disabled || !global.user}>
-          Submit
-        </Button>
-        {!global.user && <a href={'/login'}>Login to add a comment</a>}
-      </Form.Item>
-    </Form>
-  );
+  // return (
+  //   <Form
+  //   form={form}
+  //     {...layout}
+  //     name="basic"
+  //     initialValues={{}}
+  //     onFinish={onFinish}
+  //     onFinishFailed={onFinishFailed}
+  //     style={{paddingTop: '50px', display: props.display}}
+  //   >
+  //     <Form.Item name="comment" rules={[]}>
+  //       <Input.TextArea />
+  //     </Form.Item>
+  //     <Form.Item {...tailLayout}>
+  //       <Button type="secondary" onClick={()=> form.resetFields()}>
+  //         Cancel
+  //       </Button>
+  //       <Button type="primary" htmlType="submit" disabled={props.disabled || !global.user}>
+  //         Submit
+  //       </Button>
+  //       {!global.user && <a href={'/login'}>Login to add a comment</a>}
+  //     </Form.Item>
+  //   </Form>
+  // );
+  return {}
 };
 
 export default NewCommentForm
