+++ src/Components/Project/ProjectSummary.js
@@ -12,16 +12,16 @@ const ProjectSummary = (props) => {
 
   return (
     <div
-      onClick={() => projectClicked(history, project.id)}
-      className={"project-summary"}
-      style={props.style}
+      className="event"
+      // onClick={() => projectClicked(history, project.id)}
+      // className={"project-summary"}
+      // style={props.style}
     >
       <h3 className={"centered-text"}>
         <b>{project.name}</b>
       </h3>
       <p> {project.description} </p>
       <TechStack techStackAsString={stack}></TechStack>
-      <br />
     </div>
   );
 };
