+++ src/Components/index.js
@@ -1,7 +1,7 @@
 import Header from "./Header/Header";
 import Footer from "./Footer/Footer";
 import Body from "./Body";
-import Login from "./Pages/Login";
+import Login from "./Login";
 import Events from "./Events";
 import Tutorials from "./Tutorial/Tutorial";
